+++ lib/site_prism_vcr/patches/element.rb
@@ -1,0 +1,4 @@
+require_relative '../mixins/element'
+
+SitePrism::Page.extend(SPV::Mixins::Element)
+SitePrism::Section.extend(SPV::Mixins::Element)